Try to Resolve Some Problems
Following the giant spam session Prime decided to embark on yesterday,
I've been working on fixing that. The problem appears to be several
processes refuse to terminate with the rest of the program, namely
the MIND processes and something else I can't determine. One issue is
that Flask reloads the entire code when the debugger is on. I've resolved
this by setting use_reloader to False when the Flask app is run. Still,
the MIND process refuses to die with the rest. Since the MIND process is
what does sign ups and replies, that seems like the issue.

Moreover, before I realized Prime was spamming people, I was constantly
restarting it because I was having issues with the database and
groups.json. I expect that these issues were also caused by the presence
of the zombie MIND processes.

The solution that I'm currently examining is making both MIND and BODY
processes daemonic, so that they die automatically when the main process
dies. The issue that I'm now running into is that daemonic processes need
a parent process or they die immediately. Creating a SOUL process that
spawns them was my initial attempt, but that gives me an error which claims
that SOUL "can only start a process object created by current process."
MIND and BODY are both instantiated in SOUL's __init__(), so I'm not sure
where exactly the error lies.

Signed-off-by: Eliminioa <eliminioa@gmail.com>