Implement the send message function
Finally, it's basically done. There are a couple of issues that need to be
ironed out, namely the 'all' list updating, but it works in a preliminary
form!. I tried to fix the all list bug, but haven't had luck so far.

Also need to log every sent message, but that's trivial.

Signed-off-by: Eliminioa <eliminioa@gmail.com>