Implement reset groups, fix all list bug, more
Lots of stuff in this commit. I implemented a reset groups function, so
that testing can be done with the all group. That should definitely NOT be
used once the app is in use, however. Also added a function to change the
sign up thread of the bot for each side, so that sign ups can be changed if
needed. Most importantly, I fixed the annoying bug where the 'all' list
updated wrong. It was a combination of what amounted to several typos and
an oversight, most annoying indeed. Finally, made it so that the home page
displays whether the status of the MIND process is working, so that the
users can alert me if it breaks. In theory they should be able to restart
it if it fails in most cases. Maybe add a page displaying the error that
occured?

Anyway, it should be basically ready to put into full production.

Signed-off-by: Eliminioa <eliminioa@gmail.com>