Begin altering app to work with heroku
In essence, I need to eliminate soul.py and have Heroku use body.py as a
web gyro, and mind.py as a worker gyro. Also need to change to PostgreSQL,
and make config stuff environment variables. Need to add another table or
DB to facilitate that, since some config stuff needs to change, and that
kind of stuff can go in a DB.

Signed-off-by: Eliminioa <eliminioa@gmail.com>